+++ NEWS.md
@@ -1,3 +1,6 @@
+* 2021-11-15, Version 1.5.0
+  * Bug fix in the processing of the remaining time
+  * Error message if Full Disk Access is not granted
 * 2021-10-26, Version 1.4.0
   * Fix the processing of tmutil lastbackup and listbackups
 * 2021-04-20, Version 1.3.0