+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
-Fix the processing of tmutil output
+* Bug fix in the processing of the remaining time
+* Error message if Full Disk Access is not granted